move tests to more appropriate folder, avoiding helper duplication
